do_kernel_build.sh: prepare for jenkins system

+++ scripts/do_kernel_build.sh
@@ -22,28 +22,45 @@ if [ ! -z "${1}" ]; then
 	ZIPLOG="${LG_OUT_DIRECTORY}/buildzip-${KERNEL_DEVMODEL_LOWER}.log"
 	ZIPLOGD="${LG_OUT_DIRECTORY}/buildzip_details-${KERNEL_DEVMODEL_LOWER}.log"
 
-	echo "* Cleaning old ${1} log files..."
-	rm -f ${LG_OUT_DIRECTORY}/*-${KERNEL_DEVMODEL_LOWER}.log
+	if [ -z "${JENKINS_HOME}" ]; then
+		echo "* Cleaning old ${1} log files..."
+		rm -f ${LG_OUT_DIRECTORY}/*-${KERNEL_DEVMODEL_LOWER}.log
+	fi
 
-	echo "* Building ${1} kernel (log in ${KERNLOG})"
-	.zefie/scripts/cleanbuild.sh > "${KERNLOG}" 2>&1
+	if [ -z "${JENKINS_HOME}" ]; then
+		echo "* Building ${1} kernel (log in ${KERNLOG})"
+		.zefie/scripts/cleanbuild.sh > "${KERNLOG}" 2>&1
+	else
+		echo "* Building ${1} kernel"
+		.zefie/scripts/cleanbuild.sh
+	fi
 	RC=$?
 	if [ $RC -ne 0 ]; then
 		echo "Error. Please check log. Code: $RC"
 		exit $RC
 	fi
 
-	echo "* Building ${1} zip (log in ${ZIPLOG})"
-	.zefie/scripts/buildzip.sh > "${ZIPLOG}" 2>&1
+	if [ -z "${JENKINS_HOME}" ]; then
+		echo "* Building ${1} zip (log in ${ZIPLOG})"
+		.zefie/scripts/buildzip.sh > "${ZIPLOG}" 2>&1
+	else
+		echo "* Building ${1} zip"
+		.zefie/scripts/buildzip.sh
+	fi
 	RC=$?
 	if [ $RC -ne 0 ]; then
 		echo "Error. Please check log. Code: $RC"
 		exit $RC
 	fi
 
-	cp build/out/buildzip.log "${ZIPLOGD}"
-	ZIPNAME=$(tail -n3 "${ZIPLOGD}" | grep Generated | cut -d' ' -f2 | rev | cut -d'/' -f1 | rev)
-	cp "build/out/${ZIPNAME}" "${LG_OUT_DIRECTORY}"/
+	if [ -z "${JENKINS_HOME}" ]; then
+		cp build/out/buildzip.log "${ZIPLOGD}"
+		ZIPNAME=$(tail -n3 "${ZIPLOGD}" | grep Generated | cut -d' ' -f2 | rev | cut -d'/' -f1 | rev)
+		cp "build/out/${ZIPNAME}" "${LG_OUT_DIRECTORY}"/
+	else
+		ZIPNAME=$(tail -n3 "${ZIPLOGD}" | grep Generated | cut -d' ' -f2 | rev | cut -d'/' -f1 | rev)
+		cp "build/out/${ZIPNAME}" "${WORKSPACE}"/
+	fi
 else
 	echo "Usage: ${0} MODEL [conf]"
 	exit 1
